+++ app/components/AppHeader.tsx
@@ -1,21 +1,7 @@
-"use client";
-
-import { useAppNavigation } from "@/app/hooks/useAppNavigation";
 import { ShoppingCart } from "lucide-react";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
-import { MouseEvent } from "react";
 
 const AppHeader = () => {
-  const { navigateTo } = useAppNavigation();
-  const router = useRouter();
-
-  const handleCartClick = (e: MouseEvent<HTMLAnchorElement>) => {
-    e.preventDefault();
-    navigateTo("cart");
-    router.push("/cart");
-  };
-
   return (
     <header className="bg-white px-4 py-3 flex-shrink-0 z-50">
       <div className="flex items-center justify-between">
@@ -26,7 +12,6 @@ const AppHeader = () => {
           href="/cart"
           className="flex items-center space-x-1 p-2 rounded-lg"
           aria-label="장바구니"
-          onClick={handleCartClick}
         >
           <ShoppingCart className="w-5 h-5 text-gray-700" />
         </Link>
