+++ app/components/PersonalizedProduct.tsx
@@ -1,7 +1,74 @@
 import { fetchProductPreviewInfo } from "@/lib/api/productApi";
 import { fetchCustomers } from "@/lib/api/customerApi";
+import { ProductPreviewInfo } from "@/lib/types/productType";
+import { formatPriceToKor } from "@/lib/utils/constant";
+import Image from "next/image";
+import Link from "next/link";
 import { Suspense } from "react";
-import PersonalizedProductCard from "@/app/components/PersonalizedProductCard";
+
+const ProductImage = ({ product }: { product: ProductPreviewInfo }) => (
+  <div className="relative w-full aspect-square mb-2 overflow-hidden rounded-lg">
+    <Image
+      src={product.thumbnailImage ?? ""}
+      alt={product.name}
+      fill
+      className="object-cover"
+      priority
+    />
+  </div>
+);
+
+const ProductPrice = ({ product }: { product: ProductPreviewInfo }) => {
+  const { basePrice, discountedPrice } = product;
+  if (!discountedPrice) {
+    return (
+      <div>
+        <span className="text-lg font-bold text-gray-900">
+          {formatPriceToKor(basePrice)}원
+        </span>
+      </div>
+    );
+  }
+  return (
+    <div>
+      <span className="text-lg font-bold text-gray-900">
+        {formatPriceToKor(discountedPrice)}원
+      </span>
+      <span className="text-sm text-gray-500 line-through ml-1">
+        {formatPriceToKor(basePrice)}원
+      </span>
+    </div>
+  );
+};
+
+const ProductRating = ({ product }: { product: ProductPreviewInfo }) => (
+  <div>
+    <span className="text-yellow-400">★</span>
+    <span>{product.averageRating}</span>
+    <span>({product.reviewCount})</span>
+  </div>
+);
+
+const PersonalizedProductCard = ({
+  product,
+}: {
+  product: ProductPreviewInfo;
+}) => (
+  <Link key={product.id} href={`/product/${product.id}`}>
+    <ProductImage product={product} />
+    <div className="space-y-1">
+      <h3 className="text-sm font-medium text-gray-900 line-clamp-2">
+        {product.name}
+      </h3>
+      <div className="flex items-center space-x-2">
+        <ProductPrice product={product} />
+      </div>
+      <div className="flex items-center text-xs text-gray-500 gap-1">
+        <ProductRating product={product} />
+      </div>
+    </div>
+  </Link>
+);
 
 const PersonalizedProductsContent = async () => {
   const [products, customers] = await Promise.all([
@@ -21,12 +88,11 @@ const PersonalizedProductsContent = async () => {
     </section>
   );
 };
-const PersonalizedProduct = () => {
-  return (
-    <Suspense fallback={<span>로딩 중...</span>}>
-      <PersonalizedProductsContent />
-    </Suspense>
-  );
-};
+
+const PersonalizedProduct = () => (
+  <Suspense fallback={<span>로딩 중...</span>}>
+    <PersonalizedProductsContent />
+  </Suspense>
+);
 
 export default PersonalizedProduct;
